Compile without -std=c++11, not really needed
Might be a bit unconventional, downgrading the required standard, but
without these extra , and the use of isinf this compiles and runs on
Solaris 10 still which seems like a good fit for Dillo.

Additional fixes for non C++11 in here too courtesy Sevan Janiyan,
making Dillo compile and run on OS X 10.4 PowerPC with GCC 4.0.1 & 8.5.